I got sick of it, the setting of paths seemed unfriendly on a fresh install, and the settings seemed illogical. It will now boot up no matter what and always create a wiiflow.ini if there isnt one.  No problems booting from all SD, there is nothing to edit.  It also boots faster this way.
*wiiflow will now detect your boot.dol and look in the same folder for wiiflow.ini, if it is not found there it will not look anywhere else, it will be created there. its that simple.

*language.ini and titles.ini will be read also from the same folder as wiiflow.  It will not look anywhere else.

*Data directories will be created on the same drive your boot.dol is located in drive:/wiiflow, to have boot.dol on one drive (say SD) and data folders on another (USB) set data_on_usb=yes.

THE ONLY TIME you need to modify something is if you want data and dol on seperate drives.

To make your setup work with this just put your ini's in apps/wiiflow and your data dirs in  drive:/wiiflow
and set the paths in wiiflow.ini. I know this seems stupid to change this but we need static paths and it was creating a massacre on the updater.

git-svn-id: https://wiiflow-advanced.googlecode.com/svn/trunk@91 43c7226f-5572-e80b-0314-b0280708705f