*Revert some changes of last commit *Use Mem2_alloc/realloc for everything we can until I can add back a mem1lo managed memory region, 10MB of unused memmory are below the executable. *Some font changes in hopes of fixing the font bug.  It is now a member of CMenu instead of a global variable. *Make music fade in on wiifow launch, instead of immediately blasting in on launch. *Add dimok's app_booter_dol and source (thanks dimok).  Saves 99KB of dol size.
NOTE: Sometimes the wait screen will have a green flash when booting a game with alt ios.  This happens if wait thread stack or the pictures themselves get allocated into mem2 which is trashed on ios reload.  Will be fixed with the next memory commit hopefully.  Sad that i need to add some of our original mem1 management back to wiiflow.

git-svn-id: https://wiiflow-advanced.googlecode.com/svn/trunk@397 43c7226f-5572-e80b-0314-b0280708705f